added the enum files and fixed the issue of changing the type of length to double, the change has also been updated to the DB

+++ swe400_01_SingleTable/src/swe400_01_SingleTable/Fastener.java
@@ -7,5 +7,5 @@ package swe400_01_SingleTable;
  */
 public abstract class Fastener extends InventoryItem
 {
-	protected long length;
+	protected double length;
 }
